project 기능 구현 완료

+++ src/main/java/zerobase/matching/project/repository/ProjectRepository.java
@@ -1,6 +1,5 @@
 package zerobase.matching.project.repository;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import zerobase.matching.project.domain.Project;
 
 @Repository
-public interface ProjectRepository extends JpaRepository<Project, BigInteger> {
+public interface ProjectRepository extends JpaRepository<Project, Long> {
   List<Project> findAllByDueDateBefore(LocalDate date);
 }
